requirements.txt
langchain

pypdf

python-dotenv

langchain-community

langchain-openai

langchain-google-genai

neo4j

faiss-cpu
pydantic

fastapi[all]

pytest
python-json-logger
beautifulsoup4
requests
gremlinpython
boto3